requirements.txt
# Web framework
Flask>=2.3,<4
flask-cors>=4.0,<5

# Scheduling
schedule>=1.2,<2

# Audio playback
pygame>=2.5,<3

# YAML support
PyYAML>=6.0,<7

# Testing
pytest>=7.0,<9
pytest-cov>=4.0,<6

